Update EnergyData.cs
converted from float to double as it was agreed upon

+++ Heatington/Optimizer/EnergyData.cs
@@ -8,10 +8,10 @@ public struct EnergyData
     [Index(1)]
     public DateTime EndTime { get; set; }
     [Index(2)]
-    public float HeatDemandMwh { get; set; }
+    public double HeatDemandMwh { get; set; }
     [Index(3)]
-    public float ElectricityPrice { get; set; }
-    
+    public double ElectricityPrice { get; set; }
+
     public override string ToString()
     {
         string s = string.Concat(StartTime," ", EndTime," ",HeatDemandMwh," ",ElectricityPrice);
